Hasta ahora funciona todo

+++ Catalogo-Comercio/Dominio/Articulo.cs
@@ -12,7 +12,7 @@ namespace Dominio
         public string Codigo { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
-        public int Precio { get; set; }
+        public Decimal Precio { get; set; }
         public string ImagenUrl { get; set; }///URL de imagen 
         public Marcas Marca { get; set; }
         public Categorias Categoria { get; set; }
